remove appId from route definitions, tnx @berikv
